Decouple non-`blackbox`-related classes from the `blackbox` package
