--- testutil-server/src/main/java/io/spine/testing/server/blackbox/verify/count/VerifyingCounter.java
@@ -1,61 +1,0 @@
-/*
- * Copyright 2019, TeamDev. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.testing.server.blackbox.verify.count;
-
-import com.google.common.annotations.VisibleForTesting;
-
-import static com.google.common.truth.Truth.assertThat;
-
-/**
- * A counter that can verify the stored value against some {@code expected}.
- */
-@VisibleForTesting
-public final class VerifyingCounter {
-
-    private int actual;
-
-    public VerifyingCounter() {
-    }
-
-    public VerifyingCounter(int actual) {
-        this.actual = actual;
-    }
-
-    public void incrementActual() {
-        actual++;
-    }
-
-    public void incrementActual(int num) {
-        actual += num;
-    }
-
-    public void decrementActual() {
-        actual--;
-    }
-
-    public void decrementActual(int num) {
-        actual -= num;
-    }
-
-    public void verifyEquals(int expected) {
-        assertThat(actual).isEqualTo(expected);
-    }
-}
